requirements.txt
streamlit>=1.33.0
pandas>=2.2.2
numpy>=1.26.4
matplotlib>=3.8.4
seaborn>=0.13.2
scikit-learn==1.3.2
transformers>=4.40.0
nltk>=3.8.1
docx2txt>=0.8
Pillow>=10.3.0
face_recognition>=1.3.0
opencv-python>=4.9.0.80
joblib>=1.4.0
pickle-mixin>=1.0.2
